% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brier.R
\name{brier}
\alias{brier}
\alias{BrierScore}
\title{Brier score for multinomial data.}
\usage{
brier(p, labels, indices = l2i(p, labels), indicator.matrix = i2im(p,
  indices), na.rm = T)
}
\arguments{
\item{p}{An n x K matrix of probabilities, where n is the number of observations,
and K the number of mutually exclusive outcome categories.}

\item{labels}{Vector of length n, containing the labels (character or factor) of
the observed outcome categories. Ignored if \code{indices} or \code{indicator.matrix} is specified.}

\item{indices}{Optional. A vector of length n, containing the indices k, k = 1,...,K,
of the observed outcome categories. Overrides \code{labels}. Ignored if
\code{indicator.matrix} is specified.}

\item{indicator.matrix}{Optional. An n x K matrix indicating the outcome category of each observation,
where n is the number of observations, and K the number of mutually exclusive outcome categories.
Overrides \code{labels} and \code{indices}.}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
\code{brier} provides the Brier score, a vector of length 1.
}
\description{
Brier score for multinomial data. Requires the predicted (or fitted) probability matrix \code{p},
and one of the following: \code{labels}, \code{indices} or \code{indicator.matrix}.
}
\examples{
brier(matrix(1/3, nrow = 3, ncol = 3), indices = c(1,2,3))
brier(t(matrix(c(1/2, 1/4, 1/4,
               1/8, 5/8, 2/8,
               1/8, 1/8, 3/4),
               nrow = 3, ncol = 3)),
indicator.matrix = matrix(c(1,0,0,0,1,0,0,0,1), nrow = 3, byrow = TRUE))
}
