% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2.R
\name{mll}
\alias{mll}
\alias{MultinomialLogLikelihood}
\title{mll}
\usage{
mll(p, labels, indices = l2i(p, labels), na.rm = T)
}
\arguments{
\item{p}{An n x K matrix of probabilities, where n is the number of observations,
and K the number of mutually exclusive outcome categories.}

\item{labels}{Vector of length n, containing the labels (character or factor) of
the observed outcome categories. Ignored if \code{indices} is specified.}

\item{indices}{A vector of length n, containing the indices k, k = 1,...,K, of the observed outcome categories.}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
\code{mll} provides the multinomial log-likelihood.
}
\description{
Compute the multinomial log-likelihood
}
\details{
Compute the multinomial log-likelihood, for predicted or fitted values.
}
\examples{
When observing 3 outcomes with indices 1, 2 and 3,
we can obtain the log-likelihood of the null-model:
by letting all observed probabilities equal 1/3:
mll(matrix(1/3, nrow = 3, ncol = 3), indices = c(1,2,3))

}
