% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mPerformance.R
\name{mModelPerformance}
\alias{mModelPerformance}
\title{Performance of a multinomial prediction model}
\usage{
mModelPerformance(fit, ...)
}
\arguments{
\item{fit}{A model fit object.}

\item{...}{...}
}
\value{
An object of class "mPerformance".A list containing several types of performance measures: discrimination (M-index or multiclass AUC),
overall performance (Brier score and various R squares: Cox, Nagelkerke and McFadden) and calibration ().
}
\description{
Apparent (= within sample) predictive performance of a multinomial prediction model.
Works with the \code{mlogit} and \code{mnlogit} packages.
}
\references{
Nagelkerke NJ. A note on a general definition of the coefficient of determination.
Biometrika. 1991 Sep 1;78(3):691-2.
McFadden D. Conditional logit analysis of qualitative choice behavior.

Brier GW. Verification of forecasts expressed in terms of probability. Monthly weather
review. 1950 Jan;78(1):1-3.

Hand DJ, Till RJ. A simple generalisation of the area under the ROC curve
for multiple class classification problems. Machine learning. 2001 Nov 1;45(2):171-86.
}
