% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mPerformance.R
\name{mPerformance}
\alias{mPerformance}
\title{Performance of multinomial predictions.}
\usage{
mPerformance(p, labels, indices, indicator.matrix, names = colnames(p),
  na.rm = T)
}
\arguments{
\item{p}{An n x K matrix of probabilities, where n is the number of observations,
and K the number of mutually exclusive outcome categories.}

\item{labels}{Vector of length n, containing the labels (character or factor) of
the observed outcome categories. If specified, must correspond with the column names
of \code{p} or with \code{names}.}

\item{indices}{Optional. A vector of length n, containing the indices k, k = 1,...,K,
of the observed outcome categories. If specified, these indices must corresond with
their respective indices in \code{p}.}

\item{indicator.matrix}{Optional. An n x K matrix indicating the outcome category of
each observation, where n is the number of observations, and K the number of mutually
exclusive outcome categories. If specified, the order of the columns should correspond
with the order of the columns of \code{p}.}

\item{names}{Optional. What are the \code{labels} to which the columns of \code{p}
should be matched? By default, the colnames of the outcome matrix \code{p}.}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
An object of class "mPerformance", containing several types of performance measures: discrimination (M-index or multiclass AUC),
overall performance (Brier score and various R squares: Cox, Nagelkerke and McFadden) and calibration ().
}
\description{
Predictive performance for predicted or fitted values. Requires the predicted
(or fitted) probability matrix \code{p}, and one of the following: \code{labels},
\code{indices} or \code{indicator.matrix}. Preferably one of the former.
}
\references{
Nagelkerke NJ. A note on a general definition of the coefficient of determination.
Biometrika. 1991 Sep 1;78(3):691-2.
McFadden D. Conditional logit analysis of qualitative choice behavior.

Brier GW. Verification of forecasts expressed in terms of probability. Monthly weather
review. 1950 Jan;78(1):1-3.

Hand DJ, Till RJ. A simple generalisation of the area under the ROC curve
for multiple class classification problems. Machine learning. 2001 Nov 1;45(2):171-86.
}
