% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2.R
\name{mr2}
\alias{mr2}
\alias{MultinomialRSquare}
\title{Multinomial R-square}
\usage{
mr2(p, labels, indices, indicator.matrix, names = colnames(p), na.rm = T)
}
\arguments{
\item{p}{An n x K matrix of probabilities, where n is the number of observations,
and K the number of mutually exclusive outcome categories.}

\item{labels}{Vector of length n, containing the labels (character or factor) of
the observed outcome categories. If specified, must correspond with the column names
of \code{p} or with \code{names}.}

\item{indices}{Optional. A vector of length n, containing the indices k, k = 1,...,K,
of the observed outcome categories. If specified, these indices must corresond with
their respective indices in \code{p}.}

\item{indicator.matrix}{Optional. An n x K matrix indicating the outcome category of
each observation, where n is the number of observations, and K the number of mutually
exclusive outcome categories. If specified, the order of the columns should correspond
with the order of the columns of \code{p}.}

\item{names}{Optional. What are the \code{labels} to which the columns of \code{p}
should be matched? By default, the colnames of the outcome matrix \code{p}.}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
\code{mr2} provides a data.frame of R-square values by the methods of Cox, Nagelkerke and Mcfadden.
}
\description{
Compute various R-square measures: Cox, Nagelkerke and McFadden. Requires the predicted (or fitted) probability matrix \code{p},
and one of the following: \code{labels}, \code{indices} or \code{indicator.matrix}. Preferably one of the two former.
}
\examples{
# If we observe outcomes A, B and C:
labels <- c("A", "B", "c")
# The fitted probabilities of an intercept only model are given by 1/3:
probabilities <- matrix(1/3, nrow = 3, ncol = 3)
colnames(probabilities) <- labels
# Then the multinomial R-squares can be obtained with:
mr2(probabilities, labels)
# Or:
mr2(probabilities, as.factor(labels))
# Similary, we can use the indices of the observed outcome
# categories:
mr2(probabilities, indices = c(1,2,3))
}
\references{
Nagelkerke NJ. A note on a general definition of the coefficient of determination.
Biometrika. 1991 Sep 1;78(3):691-2.
McFadden D. Conditional logit analysis of qualitative choice behavior.
}
