% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mIndex.R
\name{mIndex}
\alias{mIndex}
\alias{M-index}
\alias{MulticlassAUC}
\alias{mPairwiseC}
\title{M-index for multinomial outcome data}
\usage{
mIndex(p, labels, indices = l2i(p, labels), pairwiseC, names = colnames(p))
}
\arguments{
\item{p}{An n x K matrix of probabilities, where n is the number of observations,
and K the number of mutually exclusive outcome categories.}

\item{labels}{Vector of length n, containing the labels (character or factor) of
the observed outcome categories. Ignored if \code{indices} is specified.}

\item{indices}{Optional. A vector of length n, containing the indices k, k = 1,...,K,
of the observed outcome categories. Overrides \code{labels}.}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
\code{mIndex} provides the M-index (= Multiclass AUC), a vector of length 1.

\code{mPairwiseC} provides a matrix with pairwise c-statistics. Contains both
c(i|i) and c(j|i), and the diagonal is \code{NA}.
}
\description{
M-index for multinomial data. Requires the predicted (or fitted) probability matrix \code{p},
and one of the following: \code{labels} or \code{indices}.
}
\examples{
mIndex(matrix(1/3, nrow = 3, ncol = 3), indices = 1:3)
mPairwiseC(matrix(1/3, nrow = 3, ncol = 3), indices = 1:3)
}
