% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2.R
\name{mr2}
\alias{mr2}
\alias{MultinomialRSquare}
\alias{Compute}
\alias{various}
\alias{R-square}
\alias{measures:}
\alias{Cox,}
\alias{Nagelkerke}
\alias{and}
\alias{McFadden}
\alias{R-square.}
\title{mr2}
\usage{
mr2(p, labels, indices = l2i(p, labels), na.rm = T)
}
\arguments{
\item{p}{An n x K matrix of probabilities, where n is the number of observations,
and K the number of mutually exclusive outcome categories.}

\item{labels}{Vector of length n, containing the labels (character or factor) of
the observed outcome categories. Ignored if \code{indices} is specified.}

\item{indices}{A vector of length n, containing the indices k, k = 1,...,K, of the observed outcome categories.}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
\code{mr2} provides a data.frame of R-square values by the methods of Cox, Nagelkerke and Mcfadden.
}
\description{
mr2
}
\examples{
When we observe 3 outcomes with indices 1, 2 and 3,
we can obtain the R^2 values of the null-model
by letting all observed probabilities equal 1/3:
mr2(matrix(1/3, nrow = 3, ncol = 3), indices = c(1,2,3))

}
